+++ paper/src/main/kotlin/dev/nyon/magnetic/Config.kt
@@ -21,7 +21,8 @@ data class Config(
     var expAllowed: Boolean = true,
     var itemsAllowed: Boolean = true,
     var needPermission: Boolean = false,
-    var ignoreKilledEntities: List<Identifier> = listOf()
+    var ignoreKilledEntities: List<Identifier> = listOf(),
+    var ignoreRangedWeapons: Boolean = true
 )
 
 internal fun migrate(jsonElement: JsonElement, version: Int?): Config? {