+++ changelog.md
@@ -2,7 +2,11 @@
 
 - add option to require a player to have a permission ('magnetic.ability.use') to use the ability
     - on Fabric the fabric-permissions-api has to be installed for this to work
+    - default: false
 - add option to ignore drops from entities when killed
     - you can use identifiers and tags to define these entities
     - the option is called 'ignoreKilledEntities' and can be configured via the config file or in-game in the config
       screen
+    - default: empty
+- add option to ignore drops produced by ranged weapons
+  - default: true